Improved temporary hack to make sample tests

+++ src/tests/reachability/search_algorithm_test.rs
@@ -1,13 +1,14 @@
 #[cfg(test)]
 mod reachability_search_algorithm_test{
     use test_case::test_case;
+    use crate::System::reachability::Path;
     use crate::tests::refinement::Helper::json_run_query;
     use crate::QueryResult;
     const PATH: &str = "samples/json/EcdarUniversity";
     const PATH2: &str = "samples/json/AutomatonTestReachability";
 	use std::fs::{File, self, OpenOptions};
     use std::io::prelude::*;
-
+    use std::path::Path as PPath;
     #[test_case(PATH, "reachability: Machine -> [L5](y<6); [L4](y<=6)", true; "Existing states and with right clocks")]
     #[test_case(PATH, "reachability: Machine -> [L5](); [L4](y>7)", false; "Exisiting locations but not possible with the clocks")]
     #[test_case(PATH, "reachability: Machine -> [L4](y<=6); [L5](y>=4)", true; "Switched the two states and with right clocks")]
@@ -37,32 +38,36 @@ mod reachability_search_algorithm_test{
 
 
     fn TEMPORARY_MISSING_DECLERATIONS_HACK(path: &str) {
-		// Add system declarations
-		let mut declarations = String::new();
-		let componentNames = fs::read_dir(path.to_string() + "/Components").unwrap();
-		declarations += "{\n\"name\": \"System Declarations\",\n\"declarations\": \"system ";
-		let mut first = true;
-		for filename in componentNames{
-			if !first{
-				declarations += ", ";
-			}
-			first = false;
-			declarations = declarations + &filename.unwrap().file_name().into_string().unwrap().replace(".json", "");
-		}
-		declarations += "\"\n}";
 
-		let mut file = File::create(path.to_owned() + "/SystemDeclarations.json").unwrap();
-		file.write_all(declarations.as_bytes()).unwrap();
+        if !PPath::new(&(path.to_owned() + "/SystemDeclarations.json")).exists() {
+            // Add system declarations
+            let mut declarations = String::new();
+            let componentNames = fs::read_dir(path.to_string() + "/Components").unwrap();
+            declarations += "{\n\"name\": \"System Declarations\",\n\"declarations\": \"system ";
+            let mut first = true;
+            for filename in componentNames{
+              if !first{
+                declarations += ", ";
+              }
+              first = false;
+              declarations = declarations + &filename.unwrap().file_name().into_string().unwrap().replace(".json", "");
+            }
+            declarations += "\"\n}";
 
-		// Set declarations in file
-		let componentNames = fs::read_dir(path.to_string() + "/Components").unwrap();
-		for filename in componentNames{
-			let filenamestring = &filename.unwrap().file_name().into_string().unwrap();
-			let contents = fs::read_to_string(path.to_string() + "/Components/" + filenamestring).unwrap();
-			let new = contents.replace("declarations\": \"\",", "declarations\": \"clock x, y, z;\",");
-		
-			let mut file = OpenOptions::new().write(true).truncate(true).open(path.to_string() + "/Components/" + filenamestring).unwrap();
-			file.write(new.as_bytes());		
-		}
+            let mut file = File::create(path.to_owned() + "/SystemDeclarations.json").unwrap();
+            file.write_all(declarations.as_bytes()).unwrap();
+
+            // Set declarations in file
+            let componentNames = fs::read_dir(path.to_string() + "/Components").unwrap();
+            for filename in componentNames{
+              let filenamestring = &filename.unwrap().file_name().into_string().unwrap();
+              let contents = fs::read_to_string(path.to_string() + "/Components/" + filenamestring).unwrap();
+              let new = contents.replace("declarations\": \"\",", "declarations\": \"clock x, y, z;\",");
+            
+              let mut file = OpenOptions::new().write(true).truncate(true).open(path.to_string() + "/Components/" + filenamestring).unwrap();
+              file.write(new.as_bytes());		
+            }
+        }
+
     }
 }
